Scaling and Axis Update
Axis lines no longer stick out of the grid. The scaling on the y-axis has been changed for each bound being 20 cm, thus the calculated value for the y position will be multiplied by two to compensate for the 1 px : 2 cm scale on the y-axis. The scale text at the bottom right of the scene has been removed accordingly.
The circles become chaotically apparent for high frequency and low wave speed values.
The code for changing the theme has become inconveniently large.

Suggestions
Increase the amount of circles depending on the frequency and/or wave speed inputted.
Use CSS styling for the code to change the theme.